# String.Replace(char, char) perf