Optimizes `String.Replace(char, char)`

- Main performance advantage is the widening of the initial char copy to an int. For odd numbered indexes this will include the index that differs but that is faster than avoiding it.
- When a replacement character is found there is guaranteed to be at least one iteration through the final loop so turn that around to a do while.
- Most likely to copy a character rather than replace it so cache that in a local variable. 
- Optimal pointer iterations

[Test code](https://gist.github.com/bbowyersmyth/9c13d6873b9a888f7bb7)

**When replacement char is found at the last index**

          Method |        Median |     StdDev | length |
---------------- |--------------: |-----------: |-------: |
 ClrReplace |    24.5562 ns |  0.4414 ns |      5 |
  RtReplace |    14.6848 ns |  0.2911 ns |      5 |
 NewReplace |    11.5370 ns |  0.2693 ns |      5 |
 ClrReplace |    32.0736 ns |  0.6557 ns |     12 |
  RtReplace |    25.8075 ns |  0.9885 ns |     12 |
 NewReplace |    15.5022 ns |  0.3093 ns |     12 |
 ClrReplace |    56.8245 ns |  1.3562 ns |     26 |
  RtReplace |    43.6271 ns |  0.8049 ns |     26 |
 NewReplace |    23.5174 ns |  0.3980 ns |     26 |
 ClrReplace |   143.4814 ns |  1.4644 ns |    100 |
  RtReplace |   131.2945 ns |  1.3541 ns |    100 |
 NewReplace |    88.7460 ns |  1.8378 ns |    100 |
 ClrReplace | 1,170.7136 ns | 19.3759 ns |   1000 |
  RtReplace | 1,090.3637 ns | 13.0098 ns |   1000 |
 NewReplace |   663.0800 ns |  8.2292 ns |   1000 |

**When replacement char is found at every 10th character**

          Method |      Median |     StdDev | length |
---------------- |------------: |-----------: |-------: |
 ClrReplace |   5.0385 ns |  0.1684 ns |      5 |
  RtReplace |   5.2197 ns |  0.0958 ns |      5 |
 NewReplace |   4.4108 ns |  0.0930 ns |      5 |
 ClrReplace |  30.8051 ns |  0.8720 ns |     12 |
  RtReplace |  24.4644 ns |  0.3458 ns |     12 |
 NewReplace |  18.8340 ns |  0.6120 ns |     12 |
 ClrReplace |  38.3534 ns |  1.0674 ns |     26 |
  RtReplace |  36.3841 ns |  0.5497 ns |     26 |
 NewReplace |  27.4604 ns |  0.4923 ns |     26 |
 ClrReplace | 100.4731 ns |  3.1321 ns |    100 |
  RtReplace | 116.6437 ns |  1.3477 ns |    100 |
 NewReplace |  84.7275 ns |  1.3665 ns |    100 |
 ClrReplace | 751.8988 ns | 13.7602 ns |   1000 |
  RtReplace | 961.7088 ns | 12.3669 ns |   1000 |
 NewReplace | 653.2082 ns | 13.0732 ns |   1000 |