@jkotas 2016-03-01
LGTM modulo comments. Thank you for doing this!