@bbowyersmyth 2016-03-02
Now using wstrcpy for the initial copy. Variable location modified.